+++ yolort/models/box_head.py
@@ -353,60 +353,6 @@ def _decode_pred_logits(pred_logits: Tensor):
     return boxes, scores
 
 
-class LogitsDecoder(nn.Module):
-    """
-    This is a simplified version of post-processing module, we manually remove
-    the ``torchvision::ops::nms``, and it will be used later in the procedure of
-    exporting the ONNX graph for TensorRT.
-    """
-
-    def __init__(self, strides: List[int]) -> None:
-        """
-        Args:
-            strides (List[int]): Strides of the AnchorGenerator.
-        """
-
-        super().__init__()
-        self.strides = strides
-
-    def forward(
-        self,
-        head_outputs: List[Tensor],
-        grids: List[Tensor],
-        shifts: List[Tensor],
-    ) -> Tuple[Tensor, Tensor]:
-        """
-        Just concat the predict logits, ignore the original ``torchvision::nms`` module
-        from original ``yolort.models.box_head.PostProcess``.
-
-        Args:
-            head_outputs (List[Tensor]): The predicted locations and class/object confidence,
-                shape of the element is (N, A, H, W, K).
-            anchors_tuple (Tuple[Tensor, Tensor, Tensor]):
-            grids (List[Tensor]): Anchor grids.
-            shifts (List[Tensor]): Anchor shifts.
-        """
-        batch_size = len(head_outputs[0])
-
-        all_pred_logits = _concat_pred_logits(head_outputs, grids, shifts, self.strides)
-
-        bbox_regression = []
-        pred_scores = []
-
-        for idx in range(batch_size):  # image idx, image inference
-            pred_logits = all_pred_logits[idx]
-            boxes, scores = _decode_pred_logits(pred_logits)
-            bbox_regression.append(boxes)
-            pred_scores.append(scores)
-
-        # The default boxes tensor has shape [batch_size, number_boxes, 4].
-        # This will insert a "1" dimension in the second axis, to become
-        # [batch_size, number_boxes, 1, 4], the shape that plugin/BatchedNMS expects.
-        boxes = torch.stack(bbox_regression).unsqueeze_(2)
-        scores = torch.stack(pred_scores)
-        return boxes, scores
-
-
 class PostProcess(nn.Module):
     """
     Performs Non-Maximum Suppression (NMS) on inference results
